omoves code into profileScreen
